requirements.txt
cryptography
python-jose[cryptography]
boto3
jwcrypto
python-dotenv
requests
fastapi[standard]